modify user detail for more flexibility

+++ src/main/java/com/Security/Authify/service/StudentServiceImpl.java
@@ -4,9 +4,11 @@ import com.Security.Authify.entity.StudentEntity;
 import com.Security.Authify.io.StudentMapper;
 import com.Security.Authify.io.StudentRequest;
 import com.Security.Authify.io.StudentResponse;
+import com.Security.Authify.jwtUtils.AuthUtil;
 import com.Security.Authify.repository.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,6 +21,7 @@ public class StudentServiceImpl implements StudentService{
 
     private final StudentMapper studentMapper;
     private final StudentRepository studentRepo;
+    private final AuthUtil authUtil;
 
     @Override
     public StudentResponse createStudent(StudentRequest student) {
@@ -53,6 +56,14 @@ public class StudentServiceImpl implements StudentService{
     public StudentResponse updateStudent(StudentRequest student, String stdId) {
         StudentEntity stud = studentRepo.findByStdId(stdId)
                 .orElseThrow(() -> new RuntimeException("Student not found"));
+
+        String currentUserId = authUtil.getCurrentUserId();
+        boolean isCreator = currentUserId.equals(stud.getCreatedBy());
+        boolean isAdmin = authUtil.isAdmin();
+        if(!isCreator && !isAdmin){
+            throw new AccessDeniedException("You do not have permission to perform this action");
+        }
+
         stud.setFirstName(student.getFirstName());
         stud.setLastName(student.getLastName());
         stud.setRollNo(student.getRollNo());
@@ -65,6 +76,14 @@ public class StudentServiceImpl implements StudentService{
     public void deleteStudent(String stdId) {
         StudentEntity student = studentRepo.findByStdId(stdId)
                 .orElseThrow(() -> new RuntimeException("Student not found"));
+
+        String currentUserId = authUtil.getCurrentUserId();
+        boolean isCreator = currentUserId.equals(student.getCreatedBy());
+        boolean isAdmin = authUtil.isAdmin();
+        if(!isCreator && !isAdmin){
+            throw new AccessDeniedException("You do not have permission to perform this action");
+        }
+
         studentRepo.delete(student);
         log.info("Student deleted from DB");
     }
